Move schema fetching utility functions from apps/AppList/utils to apps/utils
